Griffin_pw/requirements.txt
torch>=2.0.0
torch-audio>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
tqdm>=4.62.0
tensorboard>=2.8.0
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0
pyyaml>=6.0
einops>=0.6.0
wandb>=0.12.0
transformers>=4.20.0
datasets>=2.0.0
accelerate>=0.20.0
pytest>=6.2.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
